langchain-agent/pyproject.toml
[project]
name = "langchain-agent"
version = "0.1.0"
description = "Langchain Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "isort>=6.0.1",
    "python-dotenv>=1.1.1",
    "google-generativeai>=0.4.1",
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-community>=0.3.29",
    "wikipedia>=1.4.0",
    "numexpr>=2.11.0",
]
